Move Base64 class to org.bson.internal package
