+++ bson/src/main/org/bson/json/ShellBinaryConverter.java
@@ -18,6 +18,7 @@
 package org.bson.json;
 
 import org.bson.BsonBinary;
+import org.bson.internal.Base64;
 
 import static java.lang.String.format;
